Remove highlighted state from Menu
